# api/openapi.yaml
openapi: 3.0.3
info:
  title: Dominion Seedstars Financial Management API
  description: |
    OpenAPI 3.0 specification generated from API_DOCUMENTATION.md.
    Base URL: https://api.dominion-seedstars.com/v1
  version: 1.0.0
servers:
  - url: https://api.dominion-seedstars.com/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CashbookEntry:
      type: object
      properties:
        date:
          type: string
          format: date
        branchId:
          type: string
        branchName:
          type: string
        cashbook1:
          type: object
          properties:
            savings:
              type: integer
              description: in kobo
            loanCollection:
              type: integer
            charges:
              type: integer
            frmHO:
              type: integer
            frmBR:
              type: integer
            total:
              type: integer
        cashbook2:
          type: object
          properties:
            disbursements:
              type: integer
            expenses:
              type: integer
            total:
              type: integer
        onlineCIH:
          type: integer
        transferToSenate:
          type: integer
        netCashFlow:
          type: integer
        submittedBy:
          type: string
        status:
          type: string
          enum: [draft, submitted, approved]
...
security:
  - bearerAuth: []
paths:
...
  /cashbook/{branchId}/{date}:
    get:
      summary: Get cashbook for branch and date
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: branchId
          required: true
          schema:
            type: string
        - in: path
          name: date
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Cashbook entry
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/CashbookEntry'
  /cashbook:
    post:
      summary: Create or update daily cashbook entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashbookEntry'
      responses:
        '200':
          description: Created/Updated
...